Time: 72 ms (35.97%), Space: 13.8 MB (56.71%) - LeetHub

+++ 69-sqrtx/69-sqrtx.py
@@ -1,7 +1,7 @@
 import math
 class Solution:
     def mySqrt(self, x: int) -> int:
-        t = math.sqrt(x)//1
+        t = int(math.sqrt(x)//1)
         return int(t)
